requirements.txt
Django==1.5.1
MySQL-python==1.2.4
South==0.8.1
argparse==1.2.1
distribute==0.6.45
django-extensions==1.1.1
djangorestframework==2.3.4
ipaddr==2.1.10
requests==1.2.3
selenium==2.33.0
simplejson==3.3.0
six==1.3.0
wsgiref==0.1.2
